Rename load.py to io.py
